Added a few more C++ tests to cover the exceptions.
Made more variables and parameters const and refactored various areas to improve the readability.

Fixed a few linting errors caused by the introduction of -Wall (/W4).

Merged the coverage and debug profiles.

Improved some CMakeLists.txt files.

+++ src/hades_extensions/src/game_objects/systems/inventory.cpp
@@ -4,7 +4,7 @@
 // ----- FUNCTIONS ------------------------------
 void InventorySystem::add_item_to_inventory(const GameObjectID game_object_id, const GameObjectID item) const {
   auto inventory{get_registry()->get_component<Inventory>(game_object_id)};
-  if (inventory->items.size() == inventory->get_capacity()) {
+  if (static_cast<int>(inventory->items.size()) == inventory->get_capacity()) {
     throw InventorySpaceError(true);
   }
   inventory->items.push_back(item);
@@ -15,7 +15,7 @@ auto InventorySystem::remove_item_from_inventory(const GameObjectID game_object_
   if (inventory->items.empty()) {
     throw InventorySpaceError(false);
   }
-  if (index < 0 || index >= inventory->items.size()) {
+  if (index < 0 || index >= static_cast<int>(inventory->items.size())) {
     throw InventorySpaceError("The index is out of range.");
   }
   const int item{inventory->items[index]};
